Add support for logical, text, and list operations.
In this grammar we have four types: numerical, logical, textual, and list. But as our variables are not typed and any expression containing a variable will become untyped,
we chose to make our grammar an untyped one. Some invalid expressions (e.g. text divide by text) will generate a runtime error instead of a parse error but that's the price
we have to pay to keep our variables untyped.